+++ examples/actions/action-settle-redeem-wlp.ts
@@ -14,8 +14,8 @@ import {
   run,
   simThenMaybeExecute,
 } from "../_shared.ts";
+import { refreshOraclePrices } from "../../src/oracle/index.ts";
 import { settleRedeemWlp, updateTokenValue } from "../../src/perp/index.ts";
-import { refreshOraclePrices } from "../../src/utils/pyth.ts";
 
 run(async () => {
   const client = await buildClient();